Feat: MovieRate 평점 컬럼 Double 타입으로 변경
MovieRate 평점 컬럼 Double 타입으로 변경

+++ src/main/kotlin/com/awsmovie/controller/dto/movie/MovieRateDto.kt
@@ -5,6 +5,6 @@ import com.awsmovie.controller.dto.user.UserDto
 data class MovieRateDto(
     val user: UserDto?,
     val movieId: Long,
-    val rate: Int,
+    val rate: Double,
     val comment: String
 )
